+++ x/stake/keeper/validator.go
@@ -14,11 +14,11 @@ import (
 // get a single validator
 func (k Keeper) GetValidator(ctx sdk.Context, addr sdk.Address) (validator types.Validator, found bool) {
 	store := ctx.KVStore(k.storeKey)
-	b := store.Get(GetValidatorKey(addr))
-	if b == nil {
+	value := store.Get(GetValidatorKey(addr))
+	if value == nil {
 		return validator, false
 	}
-	k.cdc.MustUnmarshalBinary(b, &validator)
+	validator = types.MustUnmarshalValidator(k.cdc, addr, value)
 	return validator, true
 }
 
@@ -35,15 +35,13 @@ func (k Keeper) GetValidatorByPubKey(ctx sdk.Context, pubkey crypto.PubKey) (val
 // set the main record holding validator details
 func (k Keeper) SetValidator(ctx sdk.Context, validator types.Validator) {
 	store := ctx.KVStore(k.storeKey)
-	// set main store
-	bz := k.cdc.MustMarshalBinary(validator)
+	bz := types.MustMarshalValidator(k.cdc, validator)
 	store.Set(GetValidatorKey(validator.Owner), bz)
 }
 
 // validator index
 func (k Keeper) SetValidatorByPubKeyIndex(ctx sdk.Context, validator types.Validator) {
 	store := ctx.KVStore(k.storeKey)
-	// set pointer by pubkey
 	store.Set(GetValidatorByPubKeyIndexKey(validator.PubKey), validator.Owner)
 }
 
@@ -75,9 +73,8 @@ func (k Keeper) GetAllValidators(ctx sdk.Context) (validators []types.Validator)
 		if !iterator.Valid() {
 			break
 		}
-		bz := iterator.Value()
-		var validator types.Validator
-		k.cdc.MustUnmarshalBinary(bz, &validator)
+		addr := iterator.Key()[1:]
+		validator := types.MustUnmarshalValidator(k.cdc, addr, iterator.Value())
 		validators = append(validators, validator)
 		iterator.Next()
 	}
@@ -96,9 +93,8 @@ func (k Keeper) GetValidators(ctx sdk.Context, maxRetrieve int16) (validators []
 		if !iterator.Valid() || i > int(maxRetrieve-1) {
 			break
 		}
-		bz := iterator.Value()
-		var validator types.Validator
-		k.cdc.MustUnmarshalBinary(bz, &validator)
+		addr := iterator.Key()[1:]
+		validator := types.MustUnmarshalValidator(k.cdc, addr, iterator.Value())
 		validators[i] = validator
 		iterator.Next()
 	}
@@ -205,8 +201,7 @@ func (k Keeper) UpdateValidator(ctx sdk.Context, validator types.Validator) type
 
 	// always update the main list ordered by owner address before exiting
 	defer func() {
-		bz := k.cdc.MustMarshalBinary(validator)
-		store.Set(GetValidatorKey(ownerAddr), bz)
+		k.SetValidator(ctx, validator)
 	}()
 
 	// retrieve the old validator record
@@ -441,8 +436,7 @@ func (k Keeper) unbondValidator(ctx sdk.Context, validator types.Validator) type
 	k.SetPool(ctx, pool)
 
 	// save the now unbonded validator record
-	bzVal := k.cdc.MustMarshalBinary(validator)
-	store.Set(GetValidatorKey(validator.Owner), bzVal)
+	k.SetValidator(ctx, validator)
 
 	// add to accumulated changes for tendermint
 	bzABCI := k.cdc.MustMarshalBinary(validator.ABCIValidatorZero())
@@ -469,8 +463,7 @@ func (k Keeper) bondValidator(ctx sdk.Context, validator types.Validator) types.
 	k.SetPool(ctx, pool)
 
 	// save the now bonded validator record to the three referenced stores
-	bzVal := k.cdc.MustMarshalBinary(validator)
-	store.Set(GetValidatorKey(validator.Owner), bzVal)
+	k.SetValidator(ctx, validator)
 	store.Set(GetValidatorsBondedIndexKey(validator.Owner), []byte{})
 
 	// add to accumulated changes for tendermint