+++ x/stake/types/delegation.go
@@ -2,9 +2,11 @@ package types
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/wire"
 )
 
 // Delegation represents the bond with tokens held by an account.  It is
@@ -17,7 +19,54 @@ type Delegation struct {
 	Height        int64       `json:"height"` // Last height bond updated
 }
 
-// Equal returns a boolean determining if two Delegation types are identical.
+type delegationValue struct {
+	Shares sdk.Rat
+	Height int64
+}
+
+// return the delegation without fields contained within the key for the store
+func MustMarshalDelegation(cdc *wire.Codec, delegation Delegation) []byte {
+	val := delegationValue{
+		delegation.Shares,
+		delegation.Height,
+	}
+	return cdc.MustMarshalBinary(val)
+}
+
+// return the delegation without fields contained within the key for the store
+func MustUnmarshalDelegation(cdc *wire.Codec, key, value []byte) Delegation {
+	delegation, err := UnmarshalDelegation(cdc, key, value)
+	if err != nil {
+		panic(err)
+	}
+	return delegation
+}
+
+// return the delegation without fields contained within the key for the store
+func UnmarshalDelegation(cdc *wire.Codec, key, value []byte) (delegation Delegation, err error) {
+	var storeValue delegationValue
+	err = cdc.UnmarshalBinary(value, &storeValue)
+	if err != nil {
+		return
+	}
+
+	addrs := key[1:] // remove prefix bytes
+	if len(addrs) != 2*sdk.AddrLen {
+		err = errors.New("unexpected key length")
+		return
+	}
+	delAddr := sdk.Address(addrs[:sdk.AddrLen])
+	valAddr := sdk.Address(addrs[sdk.AddrLen:])
+
+	return Delegation{
+		DelegatorAddr: delAddr,
+		ValidatorAddr: valAddr,
+		Shares:        storeValue.Shares,
+		Height:        storeValue.Height,
+	}, nil
+}
+
+// nolint
 func (d Delegation) Equal(d2 Delegation) bool {
 	return bytes.Equal(d.DelegatorAddr, d2.DelegatorAddr) &&
 		bytes.Equal(d.ValidatorAddr, d2.ValidatorAddr) &&
@@ -66,8 +115,60 @@ type UnbondingDelegation struct {
 	Balance        sdk.Coin    `json:"balance"`         // atoms to receive at completion
 }
 
-// Equal returns a boolean determining if two UnbondingDelegation types are
-// identical.
+type ubdValue struct {
+	CreationHeight int64
+	MinTime        int64
+	InitialBalance sdk.Coin
+	Balance        sdk.Coin
+}
+
+// return the unbonding delegation without fields contained within the key for the store
+func MustMarshalUBD(cdc *wire.Codec, ubd UnbondingDelegation) []byte {
+	val := ubdValue{
+		ubd.CreationHeight,
+		ubd.MinTime,
+		ubd.InitialBalance,
+		ubd.Balance,
+	}
+	return cdc.MustMarshalBinary(val)
+}
+
+// unmarshal a unbonding delegation from a store key and value
+func MustUnmarshalUBD(cdc *wire.Codec, key, value []byte) UnbondingDelegation {
+	ubd, err := UnmarshalUBD(cdc, key, value)
+	if err != nil {
+		panic(err)
+	}
+	return ubd
+}
+
+// unmarshal a unbonding delegation from a store key and value
+func UnmarshalUBD(cdc *wire.Codec, key, value []byte) (ubd UnbondingDelegation, err error) {
+	var storeValue ubdValue
+	err = cdc.UnmarshalBinary(value, &storeValue)
+	if err != nil {
+		return
+	}
+
+	addrs := key[1:] // remove prefix bytes
+	if len(addrs) != 2*sdk.AddrLen {
+		err = errors.New("unexpected key length")
+		return
+	}
+	delAddr := sdk.Address(addrs[:sdk.AddrLen])
+	valAddr := sdk.Address(addrs[sdk.AddrLen:])
+
+	return UnbondingDelegation{
+		DelegatorAddr:  delAddr,
+		ValidatorAddr:  valAddr,
+		CreationHeight: storeValue.CreationHeight,
+		MinTime:        storeValue.MinTime,
+		InitialBalance: storeValue.InitialBalance,
+		Balance:        storeValue.Balance,
+	}, nil
+}
+
+// nolint
 func (d UnbondingDelegation) Equal(d2 UnbondingDelegation) bool {
 	bz1 := MsgCdc.MustMarshalBinary(&d)
 	bz2 := MsgCdc.MustMarshalBinary(&d2)
@@ -112,7 +213,68 @@ type Redelegation struct {
 	SharesDst        sdk.Rat     `json:"shares_dst"`         // amount of destination shares redelegating
 }
 
-// Equal returns a boolean determining if two Redelegation types are identical.
+type redValue struct {
+	CreationHeight int64
+	MinTime        int64
+	InitialBalance sdk.Coin
+	Balance        sdk.Coin
+	SharesSrc      sdk.Rat
+	SharesDst      sdk.Rat
+}
+
+// return the redelegation without fields contained within the key for the store
+func MustMarshalRED(cdc *wire.Codec, red Redelegation) []byte {
+	val := redValue{
+		red.CreationHeight,
+		red.MinTime,
+		red.InitialBalance,
+		red.Balance,
+		red.SharesSrc,
+		red.SharesDst,
+	}
+	return cdc.MustMarshalBinary(val)
+}
+
+// unmarshal a redelegation from a store key and value
+func MustUnmarshalRED(cdc *wire.Codec, key, value []byte) Redelegation {
+	red, err := UnmarshalRED(cdc, key, value)
+	if err != nil {
+		panic(err)
+	}
+	return red
+}
+
+// unmarshal a redelegation from a store key and value
+func UnmarshalRED(cdc *wire.Codec, key, value []byte) (red Redelegation, err error) {
+	var storeValue redValue
+	err = cdc.UnmarshalBinary(value, &storeValue)
+	if err != nil {
+		return
+	}
+
+	addrs := key[1:] // remove prefix bytes
+	if len(addrs) != 3*sdk.AddrLen {
+		err = errors.New("unexpected key length")
+		return
+	}
+	delAddr := sdk.Address(addrs[:sdk.AddrLen])
+	valSrcAddr := sdk.Address(addrs[sdk.AddrLen : 2*sdk.AddrLen])
+	valDstAddr := sdk.Address(addrs[2*sdk.AddrLen:])
+
+	return Redelegation{
+		DelegatorAddr:    delAddr,
+		ValidatorSrcAddr: valSrcAddr,
+		ValidatorDstAddr: valDstAddr,
+		CreationHeight:   storeValue.CreationHeight,
+		MinTime:          storeValue.MinTime,
+		InitialBalance:   storeValue.InitialBalance,
+		Balance:          storeValue.Balance,
+		SharesSrc:        storeValue.SharesSrc,
+		SharesDst:        storeValue.SharesDst,
+	}, nil
+}
+
+// nolint
 func (d Redelegation) Equal(d2 Redelegation) bool {
 	bz1 := MsgCdc.MustMarshalBinary(&d)
 	bz2 := MsgCdc.MustMarshalBinary(&d2)